[#398] Add trial deduplication tasks to PyBossa
We're comparing trials two by two. To understand how we create the tasks,
consider a database with trials A, B, C and D. The tasks created will be (A, B),
(B, C) and (C, D). This won't test all possible cases, because they are in the
millions with our current database. This is just an initial pass. With this
logic, we'll create NUMBER_OF_TRIALS - 1 tasks.

There's a challenge here on how to upload this to CrowdCrafting, as it only
allows 300 requests per 15 minutes. In that speed, it'll take more than 10 days
to add the ~330k tasks. Not to say that 330k tasks is already a lot.

We'll need to filter out more trials to make it feasible, specially considering
that deciding if two trials are the same isn't a trivial task.

opentrials/opentrials#398