[#505] Extract uploading to DocumentCloud from FDA_DAP processor
We want not only to upload the DocumentCloud, but also update attributes as
needed. The FDA_DAP processor already had too many responsibilities
(downloading files, merging PDFs, sending to S3), so by extracting it in a
separate processor we can not only make them run faster (as we don't need to
download files when updating data from DocumentCloud), but also improve the
overall architecture.

opentrials/opentrials#505