[#505] Only process distinct files
The problem is that a single file can be referenced by many documents, but
we're not duplicating them in DocumentCloud. This means their metadata is
incomplete. As in, if 2 documents from different drugs cite the same file,
we'll add the metadata of only one of them, and someone searching on
DocumentCloud won't know that others also cite it.

Not sure how to solve this problem. For now, we just need to be aware of it.

opentrials/opentrials#505