Dump DB schemas using psql
* Use regular Postgres commands to dump/restore the DBs

We used SQLAlchemy to do so, but it's overly complex and doesn't handle things
like indexes. This simplifies our code.

* Remove six dependency because of docker-cloud

It'll require the version it needs. No need for us to maintain it here.

* Fix tests that broke because of DB constraints

When we dumped/loaded the DB with SQLAlchemy, we weren't loading the
constraints, so tests that should be broken weren't.

* Drop all tables before restoring test databases

This way we're sure the databases contain exactly what we need and nothing else.