Update from master (#1)
* [#374] Avoid processors interruption due to errors when possible

opentrials/opentrials#374

* [#645] Convert NCT's eligibility['gender'] == 'all' to 'both'

ClinicalTrials.gov changed from using the term "Both" to using "All", which
makes us unable to import new data in our database, as we only allow genders to
be "male", "female" or "both".

Fixes opentrials/opentrials#645

* Handle NCT trials's "Unknown status"

* Dump DB schemas using psql

* Use regular Postgres commands to dump/restore the DBs

We used SQLAlchemy to do so, but it's overly complex and doesn't handle things
like indexes. This simplifies our code.

* Remove six dependency because of docker-cloud

It'll require the version it needs. No need for us to maintain it here.

* Fix tests that broke because of DB constraints

When we dumped/loaded the DB with SQLAlchemy, we weren't loading the
constraints, so tests that should be broken weren't.

* Drop all tables before restoring test databases

This way we're sure the databases contain exactly what we need and nothing else.