[#668] Pin dependencies to exact versions
This avoids our code breaking because of a change in one of our dependencies (as
just happened with `dataset==0.8.0`). I also had to set a specific version for
`datapackage` as 0.6.1 because `pip-compile` doesn't allow commits. We should be
fine, as the only changes from the version we were using to that one was
documentation and the setup.py.

opentrials/opentrials#668