[#661] Deduplicate trials via public_title
This commit implements deduplicating trial via `public_title` if we couldn't
find a trial via `identifiers`.

# Trial deduplication process

The actual process is implemented in the `helpers.find_trial()` method as:

1. Search a record (other than the actual record) with an identifier in common. If found, return its trial.
2. Search a record (other than the actual record) from a different source with the same `public_title`. If found, return its trial.
3. If not found and we're ignoring the actual record, try again including all records in our DB.

We always search in the `records` table, as we have information there on all
trials from all sources, not only the ones we consider primary.

A tricky question is why we ignore the actual `record` when searching for a
trial. Remember that we can process the same trial and record multiple times, so
they could already be in the DB. If we don't ignore at first the current record
being processed, we're guaranteed to always find a trial when updating data: the
one created from the current record. That might not be the best choice, though.
Let's see an example. Consider that our DB looks like:

* Records
| record_id | trial_id | ... |
| --------- | -------- | --- |
| A         | X        | ... |
| B         | Y        | ... |

* Trials
| trial_id | ... |
| -------- | --- |
| X        | ... |
| Y        | ... |

If we don't ignore record `A` at first, the `.find_trial()` will always return
its trial (id `X`), as it'll have the same identifiers as what we're looking
for. But what if record `B` also has the same identifiers? Doesn't matter: we
would never return its trial, as we check record `A` first. For argument's sake,
I'm assuming that the results are sorted by id, as the point is that there
exists a sorting order that results in a non-optimal deduplication (and that we
can't fix it 100% by just changing the sorting order).

To avoid this, the first pass on the deduplication algorithm ignores the current
record we're looking for. If we can't find a trial apart from the current
record, we try again considering it as well.

We also keep a log of the trial deduplication process for debugging purposes.
The `.find_trial()` method returns which method was used to find the trial, and
the trial writer writes a row in a `trial_deduplication_logs` table containing
the trial ID, record ID, deduplication method and the code's commit hash.

To get the commit hash, we needed to change how we build the Docker images.
Instead of using the automatic builds from Docker Hub, we use Travis to build
it, passing the git commit hash as the env variable `SOURCE_COMMIT`. With this,
we can access the hash using `os.environ['SOURCE_COMMIT']`.

# General improvements

There were a few improvements to the code. The biggest one was renaming
`writers.write_trial()` to `writers.write_trial_and_record()`, as we
always want to write the record as we write a trial. Because of this,
`writers.write_record()` was removed.

Fixes opentrials/opentrials#661