[#173] Extract age range (#126)
The `age_range` is a JSON in the format:

```json
{
    "min_age": "18 Years",
    "max_age": "60 Years"
}
```

The ages are always in the format of "X Units" like "1 Month", "15 Days", "3
Years". They can't be negative, for obvious reasons. It's "any" when there's
no limit, "N/A" when the limits aren't applicable, and `None` when we don't
know the limit.

Ideally, we would have parsed the ages to a better format (i.e. not text), so
we would be able to compare them. However, the information itself is unclear:
how should we put "30 Days" and "1 Month" in the same scale? Is "30 Days" equal
to "1 Month"? What about February? What about "4 Weeks"? Is it equal to "1
Month" or "28 Days"?

We will have to solve it in the API to allow querying for age ranges, but will
leave this for when we create the search filter.

Fixes opentrials/opentrials#173